SensroCharge: use long as return type to avoid casts later on

+++ src/objects/SensorCharge.cpp
@@ -35,7 +35,7 @@ unsigned int SensorCharge::getCharge() const {
     return charge_;
 }
 
-int SensorCharge::getSign() const {
+long SensorCharge::getSign() const {
     return static_cast<std::underlying_type<CarrierType>::type>(type_);
 }
